+++ lib/trailblazer/activity.rb
@@ -1,5 +1,7 @@
 module Trailblazer
   # This is DSL-independent code, focusing only on run-time.
+  #
+  # Developer's docs: https://trailblazer.to/2.1/docs/internals.html#internals-activity
   class Activity
     # include Activity::Interface # TODO
 
@@ -36,7 +38,6 @@ require "trailblazer/activity/schema/implementation"
 require "trailblazer/activity/schema/intermediate"
 require "trailblazer/activity/circuit"
 require "trailblazer/activity/circuit/task_adapter"
-require "trailblazer/activity/config"
 require "trailblazer/activity/introspect"
 require "trailblazer/activity/task_wrap/pipeline"
 require "trailblazer/activity/task_wrap/call_task"
@@ -47,5 +48,3 @@ require "trailblazer/activity/adds"
 require "trailblazer/activity/deprecate"
 require "trailblazer/option"
 require "trailblazer/context"
-
-